requirements.txt
alembic==1.12.0
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.1.10
sqlalchemy==2.0.20
fastapi[all]
pydantic==2.3.0
pydantic-settings==2.0.3
uvicorn==0.23.2
python-dotenv==1.0.0
gunicorn
pandas
redis
python-jose[cryptography]==3.3.0
passlib==1.7.4
black
isort